requirements.txt
pip
tqdm
numpy
pandas
opencv-python
bpy
bpycv
ruamel.yaml
pyntcloud